backend/requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn==0.29.0
langchain==0.2.1
langchain-community==0.2.1
langchain-openai==0.1.7
openai>=1.24.0,<2.0.0
tiktoken==0.7.0
faiss-cpu==1.12.0
python-dotenv==1.0.1
pydantic==1.10.16
requests==2.31.0

# Frontend prototype
streamlit==1.37.1

# Data generation dependencies
pyyaml==6.0.1          # YAML configuration files
faker==24.8.0          # Synthetic data generation
sentence-transformers==3.0.1 # For local embeddings

# Optional dependencies for free LLM providers
# langchain-groq==0.1.5    # For Groq API (free hosted model)
# langchain-ollama==0.1.9  # For Ollama (free local model)
# hf_xet                 # For faster model downloads from Hugging Face Xet repositories

# Utilities
httpx==0.27.0           # Async HTTP requests
tqdm==4.66.2           # Progress bars

# Optional development dependencies for code formatting and type checking
black==24.4.2
mypy==1.10.0
